HelpMeUnpack
Ok well we solved it!

But I'm a bit confused as to _why_ this is the solution.

There are two things that seem very strange in my solution.

1: I throw away 2 bytes from the decoded Vec, that are ALWAYS zero.
They come after the short and before the float.
I don't know why they are here, or why they are always zero.
They seem like some kind of padding, but its not mentioned in the problem
statement.
My best (unsupported) guess is that I was supposed to decude the original bytes
in pieces, and this padding was part of that? But I don't really think that makes sense

2: I have to interpret the bits for the float as a f32. And then CAST it to a f64
to get the solution to match what the server is expecting.
I _think_ this is a server bug, maybe the server implementation doesn't support
storing actual 32 byte floats, so its auto converted to a 64 byte float?